temp push

+++ Knikkeren/balls.cpp
@@ -57,7 +57,7 @@ Balls::Balls(const int number, const sf::Vector2f &dims, const float fps, const 
     assert(m_dims.x > 0.0f);
     assert(m_dims.y > 0.0f);
     assert(m_fps > 0.0f);
-    assert(m_div >= 2.0f);
+    assert(m_div >= 2);
 
     place_marbles(seed);
 }
@@ -72,6 +72,7 @@ void Balls::display(sf::RenderWindow &window)
 
 void Balls::moving()
 {
+    // for (int count{0}; count < )
     move();
     marbflect();
 }
@@ -169,13 +170,13 @@ sf::Color random_color(unsigned &seed)
                      frac_to_byte(min_frac + random_frac(seed)*(1.0f - min_frac)));
 }
 
-Marble random_marble(const sf::Vector2f &dims, const float frame, const float div, unsigned &seed)
+Marble random_marble(const sf::Vector2f &dims, const float frame, const int div, unsigned &seed)
 {
     assert(dims.x > 0.0f);
     assert(dims.y > 0.0f);
     assert(frame > 0.0f);
     assert(frame < 1.0f);
-    assert(div > 2.0f);
+    assert(div > 2);
 
 
     const float radius{random_radius(dims, seed)};
